backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart
pydantic>=2.0

torch
git+https://github.com/openai/CLIP.git
Pillow
numpy
scikit-learn
faiss-cpu
pandas
joblib
